move event components in to events folder
