+++ src/components/organize/events/EventList/EventListItem.tsx
@@ -2,8 +2,8 @@ import Link from 'next/link';
 import { FormattedDate, FormattedTime } from 'react-intl';
 import { ListItem, ListItemText, Typography } from '@material-ui/core';
 
-import { getNaiveDate } from '../../../utils/dateUtils';
-import { ZetkinEvent } from '../../../types/zetkin';
+import { getNaiveDate } from 'utils/dateUtils';
+import { ZetkinEvent } from 'types/zetkin';
 
 interface EventListItemProps {
     event: ZetkinEvent;
